+++ lib/src/parser/parser.dart
@@ -17,7 +17,29 @@ class DummyParser<E> extends Parser<E> {
 abstract class Parser<E> {
   bool fastParse(ParseState state);
 
-  @Deprecated("Can be removed")
+  @experimental
+  bool fastParseMany(ParseState state) {
+    while (true) {
+      if (!fastParse(state)) {
+        return true;
+      }
+    }
+  }
+
+  @experimental
+  bool fastParseMany1(ParseState state) {
+    if (!fastParse(state)) {
+      return false;
+    }
+
+    while (true) {
+      if (!fastParse(state)) {
+        return true;
+      }
+    }
+  }
+
+  @experimental
   bool fastParseSepBy(ParseState state, Parser sep) {
     if (!fastParse(state)) {
       return true;
@@ -40,9 +62,67 @@ abstract class Parser<E> {
     return true;
   }
 
+  @experimental
+  bool fastParseSkipMany1(ParseState state) {
+    if (!fastParse(state)) {
+      return false;
+    }
+
+    while (fastParse(state)) {
+      //
+    }
+
+    return true;
+  }
+
+  @experimental
+  bool fastParseSkipMany(ParseState state) {
+    while (fastParse(state)) {
+      //
+    }
+
+    return true;
+  }
+
   Tuple1<E>? parse(ParseState state);
 
-  @Deprecated("Can be removed")
+  @experimental
+  Tuple1<List<E>>? parseMany(ParseState state) {
+    final r1 = parse(state);
+    if (r1 == null) {
+      return Tuple1(<E>[]);
+    }
+
+    final list = [r1.$0];
+    while (true) {
+      final r1 = parse(state);
+      if (r1 == null) {
+        return Tuple1(list);
+      }
+
+      list.add(r1.$0);
+    }
+  }
+
+  @experimental
+  Tuple1<List<E>>? parseMany1(ParseState state) {
+    final r1 = parse(state);
+    if (r1 == null) {
+      return null;
+    }
+
+    final list = [r1.$0];
+    while (true) {
+      final r1 = parse(state);
+      if (r1 == null) {
+        return Tuple1(list);
+      }
+
+      list.add(r1.$0);
+    }
+  }
+
+  @experimental
   Tuple1<List<E>>? parseSepBy(ParseState state, Parser sep) {
     final r1 = parse(state);
     if (r1 == null) {
@@ -68,5 +148,27 @@ abstract class Parser<E> {
     }
 
     return Tuple1(list);
+  }
+
+  @experimental
+  Tuple1? parseSkipMany(ParseState state) {
+    while (fastParse(state)) {
+      //
+    }
+
+    return const Tuple1(null);
+  }
+
+  @experimental
+  Tuple1? parseSkipMany1(ParseState state) {
+    if (!fastParse(state)) {
+      return null;
+    }
+
+    while (fastParse(state)) {
+      //
+    }
+
+    return const Tuple1(null);
   }
 }